finished first round of translation from java to javascript awaiting master Wills review

+++ Sprout/Node.pde
@@ -183,7 +183,7 @@ class Node {
     }
     // Average -- divide by how many
     if (count > 0) {
-      steer.div((float)count);
+      steer.div(count);
     }
 
     // As long as the vector is greater than 0
